Decouple looksee-browser from the com.looksee.models package
The looksee-browser module is Maven-decoupled from looksee-models and
looksee-persistence, but three classes (Browser, MobileDevice, PageAlert)
lived in package com.looksee.models, creating a split-package with
looksee-models and misleading anyone pulling A11yBrowser standalone.
Move them to com.looksee.browser, update all 10 call sites, and move
the three matching tests.

Add a harness layer (com.looksee.services.browser) that converts raw
browser output into domain state objects:

  ElementStateAdapter      - both buildElementState overloads
  ImageElementStateAdapter - both buildImageElementState overloads
  PageStateAdapter         - both buildPageState overloads (@Service,
                              uses GoogleCloudStorage for screenshots)

BrowserService keeps its existing build* method signatures; each body
is now a one-line delegate into the adapter. No external callers break
because grep confirmed no code outside BrowserService.java references
these methods. getDomElementStates and the HTML extract* helpers stay
put to keep the diff scoped.
