+++ LookseeCore/looksee-browser/src/main/java/com/looksee/browser/PageAlert.java
@@ -1,4 +1,4 @@
-package com.looksee.models;
+package com.looksee.browser;
 
 
 import com.looksee.browsing.enums.AlertChoice;
