+++ LookseeCore/looksee-browser/src/main/java/com/looksee/browsing/helpers/BrowserConnectionHelper.java
@@ -3,8 +3,8 @@ package com.looksee.browsing.helpers;
 import com.looksee.browsing.BrowserFactory;
 import com.looksee.browsing.MobileFactory;
 import com.looksee.config.BrowserStackProperties;
-import com.looksee.models.Browser;
-import com.looksee.models.MobileDevice;
+import com.looksee.browser.Browser;
+import com.looksee.browser.MobileDevice;
 import com.looksee.browsing.enums.BrowserEnvironment;
 import com.looksee.browsing.enums.BrowserType;
 import io.github.resilience4j.retry.annotation.Retry;
