+++ LookseeCore/looksee-browser/src/test/java/com/looksee/browser/PageAlertTest.java
@@ -1,4 +1,4 @@
-package com.looksee.models;
+package com.looksee.browser;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
